third_party/blink/web_tests/platform/generic/http/tests/misc/frameset-disables-resource-scheduling-expected.txt



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS is_dom_content_loaded is false
PASS is_image_done is true
PASS successfullyParsed is true

TEST COMPLETE
